requirements.txt
aiogoogletrans==3.3.3
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
channels==4.2.0
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
Django==4.2.17
djangorestframework==3.15.2
exceptiongroup==1.2.2
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.13.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
